pyproject.toml
[project]
name = "dinov3-object-detection"
version = "0.1.0"
description = "Object Detection Framework using DINOv3"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.36.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=9.0.0",
    "matplotlib>=3.5.0",
    "tqdm>=4.65.0",
    "albumentations>=1.3.0",
    "pycocotools>=2.0.6",
    "tensorboard>=2.13.0",
    "scipy>=1.10.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dinov3_od"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.4",
    "black>=24.10.0",
    "isort>=5.13.2",
    "flake8>=7.1.1",
]